+++ papermerge/test/utils.py
@@ -1,6 +1,12 @@
 from django.contrib.auth import get_user_model
+from django.contrib.auth.models import (
+    Permission
+)
 
-from papermerge.core.models import Document
+from papermerge.core.models import (
+    Document,
+    Role
+)
 
 User = get_user_model()
 
@@ -26,6 +32,40 @@ def create_margaret_user():
     user.save()
 
     return user
+
+
+def create_user(username: str, perms: list) -> User:
+    """
+    Creates a user and assigns given set of permissions.
+
+    Perms is alist of permission names (as strings).
+    For example:
+
+    ['auth.view_group', 'auth.change_group']
+    """
+    user = User.objects.create_user(
+        username,
+        is_active=True
+    )
+
+    role = Role.objects.create(
+        name=f"{username}_role"
+    )
+
+    for perm_name in perms:
+        app_label, codename = perm_name.split('.')
+        perm = Permission.objects.get(
+            content_type__app_label=app_label,
+            codename=codename
+        )
+        role.permissions.add(perm)
+
+    role.save()
+    user.role = role
+    user.save()
+
+    return user
+
 
 
 def create_elizabet_user():